+++ packages/client/src/devices/InputMediaDeviceManager.ts
@@ -30,6 +30,10 @@ export abstract class InputMediaDeviceManager<
   stopOnLeave = true;
   logger: Logger;
 
+  state: T;
+
+  protected readonly call: Call;
+  protected readonly trackType: TrackType;
   protected subscriptions: Function[] = [];
   private isTrackStoppedDueToTrackEnd = false;
   private filters: MediaStreamFilterEntry[] = [];
@@ -38,11 +42,10 @@ export abstract class InputMediaDeviceManager<
     'filterRegistrationConcurrencyTag',
   );
 
-  protected constructor(
-    protected readonly call: Call,
-    public readonly state: T,
-    protected readonly trackType: TrackType,
-  ) {
+  protected constructor(call: Call, state: T, trackType: TrackType) {
+    this.call = call;
+    this.state = state;
+    this.trackType = trackType;
     this.logger = getLogger([`${TrackType[trackType].toLowerCase()} manager`]);
     if (
       deviceIds$ &&